........
  r9594 | nirhasson | 2010-01-12 22:58:05 +0000 (Tue, 12 Jan 2010) | 1 line

  RTSS: Cache path is added as FileSystem resource location when calling ShaderGenerator::setCachePath with non empty path.
........
  r9595 | sinbad | 2010-01-14 12:48:53 +0000 (Thu, 14 Jan 2010) | 1 line

  Default to not building Dx10/Dx11 rendersystems regardless, and clearly mark as experimental. To prevent people raising a hundred bugs about them not working yet (which already started)
........
  r9596 | sinbad | 2010-01-14 18:14:03 +0000 (Thu, 14 Jan 2010) | 1 line

  Patch 2922071, extended. Avoid crashes for unsupported compositors and hide them in the interface.
........
  r9597 | sinbad | 2010-01-14 18:15:15 +0000 (Thu, 14 Jan 2010) | 1 line

  OS X docs fix
........
  r9598 | sinbad | 2010-01-14 18:43:19 +0000 (Thu, 14 Jan 2010) | 1 line

  Patch 2926706: D3D10/11 shader constants fix
........
  r9599 | nirhasson | 2010-01-14 21:52:08 +0000 (Thu, 14 Jan 2010) | 2 lines

  RTSS: Shader cache path is now added to private ShaderGenerator resource group and removed when the cache path is changed.
  ShaderSystem: Sample mouse handling behavior restored to old behavior.
........
  r9600 | sinbad | 2010-01-15 11:24:12 +0000 (Fri, 15 Jan 2010) | 5 lines

  Some license clean-up:
  - A few GL files still had a GPL header (repeatedly missed in find / replaces!)
  - Plane & Quaternion Geometric Tools links should have been to Boost license
  - MovablePlane had a copy/paste credit that didn't apply
  - Removed comment reference to the main engine being LGPL
........
  r9601 | sinbad | 2010-01-15 11:31:52 +0000 (Fri, 15 Jan 2010) | 1 line

  Comment fixes
........
  r9602 | sinbad | 2010-01-15 11:34:04 +0000 (Fri, 15 Jan 2010) | 1 line

  Remove LGPL from licenses folder, no longer relevant
........
  r9603 | sinbad | 2010-01-15 17:00:52 +0000 (Fri, 15 Jan 2010) | 3 lines

  Try to resolve FreeType search ordering issues on OS X:
  - Make sure our dependencies path is before any system value in CMAKE_PREFIX_PATH
  - Prefer static library over frameworks
........
  r9604 | sinbad | 2010-01-15 18:40:16 +0000 (Fri, 15 Jan 2010) | 1 line

  Correctly locate plugin binaries from Ogre 1.6 if pointed at it
........
  r9605 | sinbad | 2010-01-15 19:20:49 +0000 (Fri, 15 Jan 2010) | 2 lines

  FindOGRE.cmake: when building against OGRE_SOURCE, include Samples/Common/include in include path
  FindOIS.h: include parent folder in include path in all cases
........
  r9606 | nirhasson | 2010-01-16 09:17:33 +0000 (Sat, 16 Jan 2010) | 2 lines

  CMake: Find Freetype changed to search from higher version to lower -
  now it OgreMain links with 2311 source if found (need to delete cache entries)
........
  r9607 | masterfalcon | 2010-01-17 03:46:22 +0000 (Sun, 17 Jan 2010) | 1 line

  Copy over some functionality from the GL rendersystem to reduce redundant texture changes.  Improves performance a little bit.
........
  r9608 | sinbad | 2010-01-17 11:26:45 +0000 (Sun, 17 Jan 2010) | 1 line

  Patch 2928231: fix Codec compile when using OGRE_STRING_USE_CUSTOM_MEMORY_ALLOCATOR
........
  r9609 | sinbad | 2010-01-17 11:33:24 +0000 (Sun, 17 Jan 2010) | 1 line

  Patch 2928233: HashedVector::mListHashDirty was not initialised correctly in a few cases
........
  r9610 | sinbad | 2010-01-17 11:38:39 +0000 (Sun, 17 Jan 2010) | 2 lines

  Patch 2928236: Null shadow textures should not be dynamic since that means after a device lost they lost their contents, and there is no handler to manage this.
  Since only get populated once, make them static instead.
........
  r9611 | sinbad | 2010-01-17 11:42:20 +0000 (Sun, 17 Jan 2010) | 1 line

  Patch 2928240: Prevent excessive logging when texture load failed in ensureloaded/ensurePrepared
........
  r9612 | sinbad | 2010-01-17 12:35:30 +0000 (Sun, 17 Jan 2010) | 1 line

  Patch 2928245: Windows 64-bit compatibility fixes
........
  r9613 | sinbad | 2010-01-17 17:33:05 +0000 (Sun, 17 Jan 2010) | 1 line

  Patch 2928242: fix problems when using logical GPU parameter indexes which end up needing more space than originally thought
........
  r9614 | sinbad | 2010-01-17 18:25:53 +0000 (Sun, 17 Jan 2010) | 3 lines

  Fix a problem with .hdr introduced by new FreeImage
  - New FreeImage includes 'hdr' twice in its format support list, once under the Radiance decoder and once under the RAW decoder. The latter doesn't work, but was overriding the native HDR decoder. Avoided this by preventing decoders overriding earlier ones (this was a memory leak too).
  - To be sure, register our custom DDS codec first, although "dds" isn't in the RAW list.
........

+++ OgreMain/src/OgreSerializer.cpp
@@ -386,7 +386,7 @@ namespace Ogre {
     {
         for(unsigned int index = 0; index < count; index++)
         {
-            flipEndian((void *)((long)pData + (index * size)), size);
+            flipEndian((void *)((size_t)pData + (index * size)), size);
         }
     }
     
@@ -395,9 +395,9 @@ namespace Ogre {
         char swapByte;
         for(unsigned int byteIndex = 0; byteIndex < size/2; byteIndex++)
         {
-            swapByte = *(char *)((long)pData + byteIndex);
-            *(char *)((long)pData + byteIndex) = *(char *)((long)pData + size - byteIndex - 1);
-            *(char *)((long)pData + size - byteIndex - 1) = swapByte;
+            swapByte = *(char *)((size_t)pData + byteIndex);
+            *(char *)((size_t)pData + byteIndex) = *(char *)((size_t)pData + size - byteIndex - 1);
+            *(char *)((size_t)pData + size - byteIndex - 1) = swapByte;
         }
     }
